pyproject.toml
[tool.poetry]
name = "stave-proj"
version = "0.1.0"
description = ""
authors = ["Saikumar Yeddula <saikumar.yeddula@v-optimai.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
jupyter = "^1.0.0"
selenium = "^4.23.1"
webdriver-manager = "^4.0.2"
python-box = "^7.2.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
